+++ data-raw/prepare_package_data.R
@@ -15,6 +15,7 @@ library(devtools)
 library(PopGenome)
 library(here)
 library(ape)
+library(readxl)
 
 source(here("data-raw/load_metadata_fn.R"))
 
@@ -878,6 +879,14 @@ use_data(btESBL_ecoli_st410_amr, overwrite = TRUE)
 ### st167
 
 
+st167_metadata_new <-
+  read_xls(
+    here("data-raw/ecoli-genomics-paper/st167-new/13073_2019_699_MOESM2_ESM.xls")
+  ) %>%
+  filter(ST == 167) %>%
+  filter(!grepl("^G", Assembly) & !grepl("_", Assembly)) %>%
+  janitor::clean_names()
+
 st167_metadata <-
   read_tsv(here("data-raw/ecoli-genomics-paper/st167/st167.tsv"))
 
@@ -891,7 +900,8 @@ st167_metadata %>%
     Country,
     `Collection Year`,
     ST
-  ) %>%  separate_rows(
+  ) %>%
+  separate_rows(
     `Data Source(Accession No.;Sequencing Platform;Sequencing Library;Insert Size;Experiment;Status)`,
     sep = ","
   ) %>%
@@ -906,7 +916,12 @@ st167_metadata %>%
     ),
     sep = ";"
   ) ->
-  btESBL_ecoli_st167_metadata
+btESBL_ecoli_st167_metadata
+
+left_join(select(st167_metadata_new, assembly),
+  btESBL_ecoli_st167_metadata,
+  by = c("assembly" = "accession")
+) -> btESBL_ecoli_st167_metadata
 
 use_data(btESBL_ecoli_st167_metadata, overwrite = TRUE)
 
@@ -914,7 +929,7 @@ use_data(btESBL_ecoli_st167_metadata, overwrite = TRUE)
 
 st167_plasm <-
   read_csv(
-    here("data-raw/ecoli-genomics-paper/st167/st167_pf_ariba_summary.csv"))
+    here("data-raw/ecoli-genomics-paper/st167-new/st167-ariba-plasmidfinder-summary.csv"))
 
 
 st167_plasm %>%
@@ -937,9 +952,9 @@ use_data(btESBL_ecoli_st167_plasmids, overwrite = TRUE)
 # st167 tree -----------------------------------------------
 
 read.tree(
-  here(
-    "data-raw/ecoli-genomics-paper/st167/clean.full.filtered_polymorphic_sites.ref_removed.snpsites.fasta.treefile")) ->
-  st167_tree
+  "data-raw/ecoli-genomics-paper/st167-new/clean.full.filtered_polymorphic_sites.fasta.treefile"
+) ->
+st167_tree
 
 st167_tree$tip.label <- gsub("_filtered","", st167_tree$tip.label)
 
@@ -949,18 +964,9 @@ use_data(btESBL_ecoli_globalst167_tree, overwrite = TRUE)
 
 # non ASC tree
 
-read.tree(
-  here("data-raw/ecoli-genomics-paper/st167/tree_no_ASC/clean.full.filtered_polymorphic_sites.fasta.treefile")
-) -> btESBL_ecoli_globalst167_tree_noASC
 
 
-midpoint.root(btESBL_ecoli_globalst167_tree_noASC) ->
-  btESBL_ecoli_globalst167_tree_noASC
 
-btESBL_ecoli_globalst167_tree_noASC$tip.label <-
-  gsub("_filtered","", btESBL_ecoli_globalst167_tree_noASC$tip.label)
-
-use_data(btESBL_ecoli_globalst167_tree_noASC, overwrite = TRUE)
 
 # overwrite non ASC trees
 
@@ -968,9 +974,7 @@ btESBL_ecoli_globalst410_tree <- btESBL_ecoli_globalst410_tree_noASC
 
 use_data(btESBL_ecoli_globalst410_tree, overwrite = TRUE)
 
-btESBL_ecoli_globalst167_tree <- btESBL_ecoli_globalst167_tree_noASC
 
-use_data(btESBL_ecoli_globalst167_tree, overwrite = TRUE)
 
 # st167 amr ---------------------------------------------------
 
@@ -978,16 +982,21 @@ use_data(btESBL_ecoli_globalst167_tree, overwrite = TRUE)
 amr.ariba167 <-
   read_csv(
     here(
-      "data-raw/ecoli-genomics-paper/st167/st167_ariba_srst2_summary.csv"))
+      "data-raw/ecoli-genomics-paper/st167-new/st167-ariba-srst2-summary.csv"
+    )
+  )
 
 amr.ariba167 %>%
-  mutate(name = gsub("\\./", "", name),
-         name = gsub("/report.tsv", "", name),
-         name = gsub("_filtered", "", name),
-         name = gsub("#","_", name)) %>%
+  mutate(
+    name = gsub("\\./", "", name),
+    name = gsub("/report.tsv", "", name),
+    name = gsub("_filtered", "", name),
+    name = gsub("#", "_", name)
+  ) %>%
   pivot_longer(-name,
-               names_to = c( "cluster", ".value"),
-               names_sep = "\\.") %>%
+    names_to = c("cluster", ".value"),
+    names_sep = "\\."
+  ) %>%
   mutate(gene = sapply(str_split(ref_seq, "__"), function(x) x[3])) %>%
   filter(match == "yes") %>%
   mutate(gene = case_when(
@@ -995,7 +1004,7 @@ amr.ariba167 %>%
     TRUE ~ gene
   )) %>%
   select(name, gene) ->
-  btESBL_ecoli_st167_amr
+btESBL_ecoli_st167_amr
 
 use_data(btESBL_ecoli_st167_amr, overwrite = TRUE)