+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unpublished
+
+### Added
+
+### Fixed
+
+### Changed
+
+- Changed st167 e coli global analysis to not use enterobase
+
 ## [1.2.0] - 2022-03-25
 ### Added
 - Longitudinal carriage model ceftriaxone vs all other abx