+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed st167 e coli global analysis to not use enterobase
+- changed st410 e coli global analysis to not use enterobase
 
 ## [1.2.0] - 2022-03-25
 ### Added